extract validNumbers out as separated function
validNumbers will be helpful in coming exercises to get completed, so it's good to pull it out and reuse the potential of code
