+++ js-exercises/is-triangle/isTriangle.js
@@ -1,28 +1,4 @@
-const validNumbers = (...values) => {
-  let result = true;
-
-  if (values === undefined || values.length === undefined || values.length <= 0) {
-    result = false;
-    return result;
-  }
-
-  for (let index = 0; index < values.length; index += 1) {
-    const value = values[index];
-    if (typeof value !== 'number') {
-      result = false;
-      break;
-    }
-  }
-
-  if (!result) {
-    throw new TypeError(`
-    One of the provided values is not a number.
-    Check the values sent - ${values.toString()}
-    `);
-  }
-
-  return result;
-};
+import { validNumbers } from '../common/validNumbers';
 
 const isTriangle = (side1, side2, side3) => {
   let isValidTriangle = false;
